chore: 拆分VDD逻辑
